+++ CarShop/CarShop.ServiceDefaults/ServiceInterfaces/CarStorage/GetCarsResult.cs
@@ -9,6 +9,6 @@ namespace CarShop.ServiceDefaults.ServiceInterfaces.CarStorage
     public class GetCarsResult
     {
         public IEnumerable<Car> Cars { get; init; }
-        public long TotalResultsCount { get; init; }
+        public int TotalResultsCount { get; init; }
     }
 }
